Installer: [FIX] mount and umount sysfs in chroot
Exec preconditions are executed inside the "system under install" after
chrooting into it. To prepare this chroot we now also mount the sysfs
and unmount it afterwards. This was neccessary because gen_initrd
mounted this filesystem and thus made umounting the chroot impossible
because the sysfs was not unmounted. Furthermore, having sysfs is a good
idea anyway.