Installer: [TEST] loop for use test
This patch changes the 00-load tests for all installer modules to the
style used in MCP already. It has a list of all modules, loops through
this list and calls "require $module; new $module" on each module. This
is easily extendable and allows plans without done_testing and still now
manual calculation.