[FIX] don't use basedir in elements of $self->images
The images installed during installation are safed in $self->images. The
elements of this list used to contain the basedir in their mount
part. Since this list is used to generate the fstab file we need a
simple way to get the mount directory without basedir. I decided the
easiest way is to save the mount dir without basedir and let unmount add
the basedir part as needed.