[RFK] changed API of Precondition::guest_install()
Precondition::guest_install() expected an image and a partition and
worked based on which of these two was set. This was hardly extensible
and so I reworked it. Now it expects a mount mode and mount options and
works on them. Installer::Base as the only user of the function is
changed to create this information and use the new API.

I also cleaned guest_install up so it is easier to see what is done for
which mount mode.