[RFK] merge precondition_install and prepared_install
precondition_install was used to generate a config for
prepared_install. This means the first function analysed the
precondition and generated a config for the second one. The second then
had to analyse the config to work on. This doesn't make any sense since
precondition_install was the only user of prepared_install.

Now there is only one function named precondition_install that decides
what to do based on the precondition and then installs the precondition
as expected. This also makes the API easier because precondition_install
is inherited by every precondition and is no longer a Base function.