no more (*::)Method::Signatures at all
The installer is naturally too difficult
to unittest to catch all the issues that
might arise. And they really happen - we
were hit by strange issues after an upgrade
of upstream libs and don't know where the
error is.

So I just dropped the dependency to
MooseX::Method::Signatures and use plain
Moose and classic methods.