try dhclient if no network
Some images seem to not start network per default. To fix this we check
whether tapper server can be reached (ping) and if not we call dhclient
to set up the network. This will not always work but since we do not
check the return value a false negative is probably not an issue.